Moved out models into separate modules
